Import code from the spip-admin project.
Import the code that was so far developed as part of another project.
Inline some functions to make it a stand-alone library.

The translation function used by optparse should be defined in the
_OPTPARSE_T constant.

Add a basic readme to show how to use the library.

Apply GPLv2 license to the code.