Small fixes for _verify_settings_dependencies
Mark an empty if clause for future correction.

The exception "callback not callable" fires off in the undesirable case
where the callback is Null. This makes verifying the validity of the
assertion for the exception thrown just before since an OptionError will
still be thrown.
Make it fire off only when the attribute callback is not Null.

Correct two exception messages: change option into action.